authentication/requirements.txt
fastapi==0.110.0
pydantic==1.10.2
python_jose==3.3.0
